Seven latent defects the scan did find, and the file with no room for the fix
None of these is a live failure. They are dead fallbacks, dead branches and
signatures that disagree with their implementation -- the kind of thing that is
wrong now and becomes a bug the day someone relies on it.

`String(raw.name || raw.name || '')` appears in `use-pod-context.ts`,
`use-datastores.ts` and `table-builder.tsx`, and `createdTable?.name ||
createdTable?.name || tableName` in the fourth. The second operand is dropped
rather than replaced: `TableResponse` and `TableSummaryResponse` only ever emit
`name`. `table_name` exists as an input alias on `CreateTableRequest` and as a
field on two unrelated types, so writing it as the fallback would have invented a
value the API never sends.

`models-settings.tsx` computed `profileStatus(profile) ?? (isHarness ? {label:
'Computer not connected'} : null)`, where `profileStatus` is non-nullable and all
three of its returns are object literals -- so that label could never render.
Deleting it loses nothing: `runtimeAvailabilityLabel` already reports the
disconnected states as Offline, Not installed and Unavailable.

`turns.ts` guarded on `record(result.answers) &&`, and `record()` returns `{}` for
any non-object, so the left operand was always true.

`FileTransactionWriter._ensure_directory_path` and
`ScheduleFilterTaskQueue.enqueue` both declared signatures their only
implementations do not have. In both cases the declaration was the wrong one, so
it moves to match. For the queue that also removes a runtime `ValueError` whose
only job was to re-check what the type now states.

`function_runtime_endpoint_cache` had a `raise AssertionError` after an
unconditional return, orphaned when a retry loop was removed in 73f1e88ae.

The one worth reading twice is `FunctionDispatcher`. Its `except BaseException`
re-raises `CancelledError` but let `KeyboardInterrupt` and `SystemExit` fall
through to `fail_dispatch`, so a process shutting down recorded runs as failed on
its way out. `agent_runner_service` guards the same boundary the same way, and
its comment records what the missing guard cost there.

That fix did not fit. `function_dispatcher.py` was at 597 lines against the
600-line ceiling `check_architecture.py` enforces, and the guard plus its
explanation took it to 605 -- so the first version of this shipped with the
reasoning compressed to a single line to get under the limit. A file that cannot
afford a comment is the gate asking for something.

So the failure taxonomy and the two functions that turn a failure into the one
line a user reads move to `dispatch_failures.py`. They are a pure function of the
exception, and classifying at the process boundary is its own job per
CONTRIBUTING. `_execution_error` and `_runtime_failure_message` stay reachable as
class attributes through the `staticmethod(...)` binding the file already used
for `_terminal_logs`, so existing call sites and tests are untouched. The
dispatcher is 560 lines, and the guard has its full comment back.
